requirements.txt
python-barcode

pip~=21.1.2
setuptools~=57.0.0